requirements.txt
streamlit==0.63.0
pandas==1.0.5
numpy==1.19.0
scikit-learn==0.23.1
Pillow==7.2.0
Pillow-PIL==0.1.dev0
